requirements.txt
Flask==3.0.3
requests==2.32.3
pytest==8.3.3
PyYAML==6.0.2
requests==2.32.3
pytest==8.3.3
